rename auth state to user state
